--- a/apps/files_external/3rdparty/icewind/smb/src/Exception/ConnectException.php
+++ b/apps/files_external/3rdparty/icewind/smb/src/Exception/ConnectException.php
@@ -1,0 +1,10 @@
+<?php
+/**
+ * Copyright (c) 2014 Robin Appelman <icewind@owncloud.com>
+ * This file is licensed under the Licensed under the MIT license:
+ * http://opensource.org/licenses/MIT
+ */
+
+namespace Icewind\SMB\Exception;
+
+class ConnectException extends Exception {}