--- a/apps/files_external/3rdparty/icewind/streams/.gitignore
+++ b/apps/files_external/3rdparty/icewind/streams/.gitignore
@@ -1,0 +1,3 @@
+.idea
+vendor
+composer.lock