New SMB storage backend based on https://github.com/icewind1991/SMB which supports either wrapping smbclient or use libsmbclient directly trough libsmbclient-php.

All tests pass locally (both with smbclient and libsmbclient)

Both methods also provide a nice speedup compared to smb4php, on my local machine I get the following times for running the smb unit tests

| old (smb4php)  | new (smbclient)   | new (libsmbclient)  |
| --------------------|-----------------------| ------------------------ |
| 16s                   | 8s                        | 1s                           |

cc @DeepDiver1975 @PVince81 @Xenopathic 

Fixes #6092, fixes #9467, fixes #9480