@scrutinizer-notifier 2015-02-05
A [new inspection](https://scrutinizer-ci.com/g/owncloud/core/inspections/d87945fa-b624-4d23-92b6-12737217937d) was created.