@scrutinizer-notifier 2015-02-05
The [inspection](https://scrutinizer-ci.com/g/owncloud/core/inspections/d87945fa-b624-4d23-92b6-12737217937d) completed: **22 new issues, 12 updated code elements**