@scrutinizer-notifier 2015-02-05
A [new inspection](https://scrutinizer-ci.com/g/owncloud/core/inspections/669d57d0-d302-4cd0-b277-f64fb221f6fb) was created.