@scrutinizer-notifier 2015-02-05
The [inspection](https://scrutinizer-ci.com/g/owncloud/core/inspections/669d57d0-d302-4cd0-b277-f64fb221f6fb) completed: **38 new issues, 12 updated code elements**