@icewind1991 2015-02-05
@PVince81 @DeepDiver1975 can we merge this once stable8 is branched off?