@icewind1991 2015-02-06
> So, do the unit tests pass now ?

We have passing unit tests for this over at bhttps://github.com/owncloud/core/issues/13922