@scrutinizer-notifier 2015-02-06
A [new inspection](https://scrutinizer-ci.com/g/owncloud/core/inspections/dd7bd6f1-5e6f-4b79-947c-f82b11ee358e) was created.