@scrutinizer-notifier 2015-02-06
The [inspection](https://scrutinizer-ci.com/g/owncloud/core/inspections/dd7bd6f1-5e6f-4b79-947c-f82b11ee358e) completed: **24 new issues, 12 updated code elements**