@PVince81 2015-02-09
CC @jnfrmarks once this is merged we need to regression test SMB mounts