@PVince81 2015-02-09
- [ ] BUG: enable encryption, create empty file in the web UI, download that file :no_entry_sign: FAIL