@PVince81 2015-02-09
- [ ] BUG: delete file when encryption is enabled:
```
{"reqId":"e61465d9445dfaee825f73d4669f06e9","remoteAddr":"127.0.0.1","app":"index","message":"Exception: {\"Message\":\"Invalid request for smb:\\\/\\\/localhost\\\/vincent\\\/temp\\\/smbmount\\\/P1050409.JPG\",\"Code\":2,\"Trace\":\"#0 \\\/srv\\\/www\\\/htdocs\\\/owncloud\\\/apps\\\/files_external\\\/3rdparty\\\/icewind\\\/smb\\\/src\\\/NativeState.php(72): Icewind\\\\SMB\\\\NativeState->handleError('smb:\\\/\\\/localhost...')\\n#1 \\\/srv\\\/www\\\/htdocs\\\/owncloud\\\/apps\\\/files_external\\\/3rdparty\\\/icewind\\\/smb\\\/src\\\/NativeState.php(146): Icewind\\\\SMB\\\\NativeState->testResult(false, 'smb:\\\/\\\/localhost...')\\n#2 \\\/srv\\\/www\\\/htdocs\\\/owncloud\\\/apps\\\/files_external\\\/3rdparty\\\/icewind\\\/smb\\\/src\\\/NativeShare.php(152): Icewind\\\\SMB\\\\NativeState->unlink('smb:\\\/\\\/localhost...')\\n#3 \\\/srv\\\/www\\\/htdocs\\\/owncloud\\\/apps\\\/files_external\\\/lib\\\/smb.php(122): Icewind\\\\SMB\\\\NativeShare->del('\\\/temp\\\/smbmount\\\/...')\\n#4 \\\/srv\\\/www\\\/htdocs\\\/owncloud\\\/apps\\\/files_trashbin\\\/lib\\\/storage.php(84): OC\\\\Files\\\\Storage\\\\SMB->unlink('P1050409.JPG')\\n#5 \\\/srv\\\/www\\\/htdocs\\\/owncloud\\\/lib\\\/private\\\/files\\\/view.php(806): OCA\\\\Files_Trashbin\\\\Storage->unlink('P1050409.JPG')\\n#6 \\\/srv\\\/www\\\/htdocs\\\/owncloud\\\/lib\\\/private\\\/files\\\/view.php(442): OC\\\\Files\\\\View->basicOperation('unlink', '\\\/folder\\\/P105040...', Array)\\n#7 \\\/srv\\\/www\\\/htdocs\\\/owncloud\\\/lib\\\/private\\\/files\\\/filesystem.php(651): OC\\\\Files\\\\View->unlink('\\\/folder\\\/P105040...')\\n#8 \\\/srv\\\/www\\\/htdocs\\\/owncloud\\\/apps\\\/files\\\/ajax\\\/delete.php(31): OC\\\\Files\\\\Filesystem::unlink('\\\/folder\\\/P105040...')\\n#9 \\\/srv\\\/www\\\/htdocs\\\/owncloud\\\/lib\\\/private\\\/route\\\/route.php(135) : runtime-created function(1): require_once('\\\/srv\\\/www\\\/htdocs...')\\n#10 [internal function]: __lambda_func(Array)\\n#11 \\\/srv\\\/www\\\/htdocs\\\/owncloud\\\/lib\\\/private\\\/route\\\/router.php(250): call_user_func('\\\\x00lambda_3371', Array)\\n#12 \\\/srv\\\/www\\\/htdocs\\\/owncloud\\\/lib\\\/base.php(763): OC\\\\Route\\\\Router->match('\\\/apps\\\/files\\\/aja...')\\n#13 \\\/srv\\\/www\\\/htdocs\\\/owncloud\\\/index.php(36): OC::handleRequest()\\n#14 {main}\",\"File\":\"\\\/srv\\\/www\\\/htdocs\\\/owncloud\\\/apps\\\/files_external\\\/3rdparty\\\/icewind\\\/smb\\\/src\\\/NativeState.php\",\"Line\":44}","level":4,"time":"2015-02-09T10:26:56+00:00","method":"POST","url":"\/owncloud\/index.php\/apps\/files\/ajax\/delete.php"}
```