@scrutinizer-notifier 2015-02-09
A [new inspection](https://scrutinizer-ci.com/g/owncloud/core/inspections/4b824e54-3c68-4a69-bd19-75c69cb6807c) was created.