@scrutinizer-notifier 2015-02-09
A [new inspection](https://scrutinizer-ci.com/g/owncloud/core/inspections/5f63fdb2-c122-49d9-a9e1-ab42ca1f6cc4) was created.