@icewind1991 2015-02-09
@PVince81 fixed both issues, thanks for testing