@scrutinizer-notifier 2015-02-09
The [inspection](https://scrutinizer-ci.com/g/owncloud/core/inspections/4b824e54-3c68-4a69-bd19-75c69cb6807c) completed: **18 new issues, 12 updated code elements**