@scrutinizer-notifier 2015-02-09
The [inspection](https://scrutinizer-ci.com/g/owncloud/core/inspections/5f63fdb2-c122-49d9-a9e1-ab42ca1f6cc4) completed: **18 new issues, 12 updated code elements**