@icewind1991 2015-02-09
@owncloud-bot retest this please