@icewind1991 2015-02-10
> I'd like to see unit tests passing against a windows share. Once done we can merge this .... finally

Ran the unit tests against a windows server 2012R2 vm