@DeepDiver1975 2015-02-10
> Ran the unit tests against a windows server 2012R2 vm

great - thx