@PVince81 2015-02-10
@icewind1991 if I remember well you said there were actually two implementations: one used when libsmbclient is available and one that falls back to smbclient. Is that still the case ?
Did you run the test for both cases ?