@scrutinizer-notifier 2015-02-11
A [new inspection](https://scrutinizer-ci.com/g/owncloud/core/inspections/b94d51af-2d69-45e0-b661-02ca44624ac0) was created.