@icewind1991 2015-02-11
@owncloud-bot retest this please