@icewind1991 2015-02-11
> Did you run the test for both cases ?

Did now, found an upstream [bug](https://bugzilla.samba.org/show_bug.cgi?id=11094) and worked around it. Tests now run for me against both windows and linux servers with both backends