@DeepDiver1975 2015-02-11
@MorrisJobke we have to respect the different operation modes in our autotest execution. Might be tricky .....