@MorrisJobke 2015-02-12
> @MorrisJobke we have to respect the different operation modes in our autotest execution. Might be tricky .....

What does that mean?