@icewind1991 2015-02-12
Testing both the libsmbclient and the smblcient implementations of the smb library