@icewind1991 2015-02-12
enabling disabling the libsmbclient extension