@DeepDiver1975 2015-02-13
I guess we should follow the mechanism in 3rdparty and remove the tests folder from icewind/smb and icewind/streams

@icewind1991 or maybe even you don't package test within the composer package?