@scrutinizer-notifier 2015-02-13
The [inspection](https://scrutinizer-ci.com/g/owncloud/core/inspections/40132c04-9a5d-45bd-a811-d75154658a35) completed: **23 new issues, 12 updated code elements**