@scrutinizer-notifier 2015-02-13
A [new inspection](https://scrutinizer-ci.com/g/owncloud/core/inspections/f91c963e-3d14-46ea-ac46-d802065f8d10) was created.