@scrutinizer-notifier 2015-02-13
The [inspection](https://scrutinizer-ci.com/g/owncloud/core/inspections/f91c963e-3d14-46ea-ac46-d802065f8d10) completed: **19 new issues, 12 updated code elements**