@scrutinizer-notifier 2015-02-16
A [new inspection](https://scrutinizer-ci.com/g/owncloud/core/inspections/abd08f3f-7d94-47f2-9303-a3edf503424e) was created.