@DeepDiver1975 2015-02-16
rebased and tested against samba server running inside docker :+1: 

issues with windows are to be seen as local network/setup issues