@MorrisJobke 2015-02-16
@owncloud-bot retest this please