@LukasReschke 2015-02-19
Please add as feature to https://github.com/owncloud/core/wiki/ownCloud-8.1-Features as well – THX.